enable more warnings
Fix some warnings

+++ spa/include/spa/pod/builder.h
@@ -563,7 +563,7 @@ spa_pod_builder_addv(struct spa_pod_builder *builder, va_list args)
 {
 	int res = 0;
 	struct spa_pod_frame *f = builder->state.frame;
-	uint32_t ftype = f ? f->pod.type : SPA_TYPE_None;
+	uint32_t ftype = f ? f->pod.type : (uint32_t)SPA_TYPE_None;
 
 	do {
 		const char *format;
