+++ spa/include/spa/pod/parser.h
@@ -445,7 +445,7 @@ do {											\
 static inline int spa_pod_parser_getv(struct spa_pod_parser *parser, va_list args)
 {
 	struct spa_pod_frame *f = parser->state.frame;
-        uint32_t ftype = f ? f->pod.type : SPA_TYPE_Struct;
+        uint32_t ftype = f ? f->pod.type : (uint32_t)SPA_TYPE_Struct;
 	const struct spa_pod_prop *prop = NULL;
 	int count = 0;
